Make sure GEP widths are correct when loading from vtable elements
We know that vtable metadata contains a bunch of pointer type width
elements, although the underlying type that's been stored in some
elements is only a usize. But we still need to make sure we treat the
pointer to the vtable as if it points to a full pointer type width
element, so that any GEPs generate the correct offset.